Finished

+++ services/cars.js
@@ -37,24 +37,42 @@ async function createCar(car) {
   await result.save();
 }
 
-async function deleteById(id) {
+async function deleteById(id, ownerId) {
+  const existing = await Car.findById(id);
+
+  
+  if (existing.owner == ownerId) {
+    return false;
+  }
+
   await Car.findByIdAndDelete(id);
+  return true;
 }
 
-async function updateById(id, car) {
+async function updateById(id, car, ownerId) {
   const existing = await Car.findById(id);
+
+  if (existing.owner != ownerId) {
+    return false;
+  }
 
   existing.name = car.name;
   existing.description = car.description;
-  existing.imageUrl = car.imageUrl;
+  existing.imageUrl = car.imageUrl || undefined;
   existing.price = car.price;
-  existing.price = car.accesories;
+  existing.accessories = car.accessories;
 
   await existing.save();
+
+  return true;
 }
 
-async function attachAccessory(carId, accessoryId) {
+async function attachAccessory(carId, accessoryId, ownerId) {
   const existing = await Car.findById(carId);
+
+  if (existing.owner != ownerId) {
+    return false;
+  }
 
   existing.accessories.push(accessoryId);
